fixes filename and folder problem
Previously files we just getting saved in the root director of our
bucket. Solved this by adding today's date as a folder and prepending
our filename with that. The next thing we did via
crypto.randomBytes is create a random string to save out files as
on AWS. The last thing we did was promisify both the s3.upload and
crypto.randomBytes so both implement promises that are resolved or
rejected based on whether or not they work.