Fixes mongoose middleware broken path
Config/database was incorrectly pathed and needed a ../.. in front
of it to access the correct directory.
Adds upload mongoose model requiring title and url as attributes
Implements this model in the bin/script so we're testing everything
from the command line -> AWS -> mongo which is all working correctly.
Removed example model because we don't need it anymore.